backend/requirements.txt
tensorflow==2.15.0
torch>=1.9.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.5.0
traci>=1.15.0
sumolib>=1.15.0
flwr==1.4.0
gym>=0.21.0
scipy>=1.9.0
pandas>=1.5.0
seaborn>=0.11.0
protobuf>=3.20.3,<4.0.0

fastapi
uvicorn
pydantic
websockets
python-dotenv
httpx
psycopg2-binary
